// packages/@azure/keyvault/lib/models/parameters.ts
export const acceptLanguage: msRest.OperationParameter = {
  parameterPath: "acceptLanguage",
  mapper: {
    serializedName: "accept-language",
    defaultValue: 'en-US',
    type: {
      name: "String"
    }
  }
};
export const apiVersion: msRest.OperationQueryParameter = {
  parameterPath: "apiVersion",
  mapper: {
    required: true,
    serializedName: "api-version",
    type: {
      name: "String"
    }
  }
};
// ...
export const maxresults: msRest.OperationQueryParameter = {
  parameterPath: [
    "options",
    "maxresults"
  ],
  mapper: {
    serializedName: "maxresults",
    constraints: {
      InclusiveMaximum: 25,
      InclusiveMinimum: 1
    },
    type: {
      name: "Number"
    }
  }
};
// ...
export const secretName0: msRest.OperationURLParameter = {
  parameterPath: "secretName",
  mapper: {
    required: true,
    serializedName: "secret-name",
    constraints: {
      Pattern: /^[0-9a-zA-Z-]+$/
    },
    type: {
      name: "String"
    }
  }
};
export const secretName1: msRest.OperationURLParameter = {
  parameterPath: "secretName",
  mapper: {
    required: true,
    serializedName: "secret-name",
    type: {
      name: "String"
    }
  }
};
export const secretVersion: msRest.OperationURLParameter = {
  parameterPath: "secretVersion",
  mapper: {
    required: true,
    serializedName: "secret-version",
    type: {
      name: "String"
    }
  }
};
// ...
export const vaultBaseUrl: msRest.OperationURLParameter = {
  parameterPath: "vaultBaseUrl",
  mapper: {
    required: true,
    serializedName: "vaultBaseUrl",
    defaultValue: '',
    type: {
      name: "String"
    }
  },
  skipEncoding: true
};